Share a device-runtime substrate across experimental CUDA solvers
De-duplicate the three experimental CUDA modules (rigid_body_state_batch_cuda,
vbd_block_descent_cuda, deformable_psd_projection_cuda) onto shared building
blocks so AVBD-CUDA, rigid-IPC GPU, and PD-IPC GPU do not each reinvent the
device plumbing (PLAN-031).

- cuda_runtime.cuh/.cpp: single isCudaRuntimeAvailable (was 3 declarations + 1
  definition + a cross-TU comment), single throwIfCudaError (was 3 definitions),
  checkLastError, and an inline launchGrid1D ceil-div helper (was hand-rolled at
  7+ sites). The definitions live in a host .cpp so callers' .cu files need not
  compile std::format.
- device_buffer.cuh: one owning DeviceBuffer<T=double> with opt-in resident
  grow-and-reuse (ensure()/allocationCount()/release()), subsuming the former
  DeviceDoubleBuffer, DeviceArray<T>, and ResidentDeviceBuffer.
- detail/rigid_integration_core.hpp: a per-body __host__ __device__ orientation
  integration core (behind a __CUDACC__-guarded DART_EXPERIMENTAL_HD macro) that
  the CPU batch kernel and the CUDA kernel both call, deleting the divergent
  __device__ integrateOrientation re-derivation (single source of truth).

Behavior change (experimental, private CUDA path): VBD's CUDA-failure path now
throws sx::InvalidOperationException with the shared "CUDA <op> failed: ..."
message instead of a local std::runtime_error with a "CUDA VBD ..." message,
unifying the error contract. Backward compatible: InvalidOperationException
derives from std::runtime_error, and no in-tree caller or test keys on the old
type/message.

Boundary/packaging unchanged: shared device code stays .cuh/.cpp under
compute/cuda/ (plus the detail/ core) in the build-only
dart-simulation-experimental-cuda static library, never installed/exported and
never in a default Pixi feature or wheel manifest.

Verified: pixi run -e cuda test-cuda (3/3 parity within the Phase-5 1.78e-15
tolerance), pixi run test-simulation-experimental (61/61), check-api-boundaries,
check-compute-backend-boundaries, check-no-gpu-runtime-dependencies, and lint.
